Add SIM, EPHIN housekeeping and CPE MSIDs

+++ make_units_cxc.py
@@ -1,30 +1,50 @@
 """
 Create the default unit system as found in the CXC telemetry FITS files.
 """
+import re
 import glob
+import argparse
 import cPickle as pickle
 import os
 import pyfits
 
-import Ska.engarchive.converters
+from Ska.engarchive.converters import _get_deahk_cols, ALIASES
 
+
+def get_options(args=None):
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--dry-run",
+                        action="store_true",
+                        help="Dry run (no actual file or database updates)")
+    return parser.parse_args(args)
+
+opt = get_options()
 
 units = {}
-dirs = glob.glob('/data/cosmos2/eng_archive/data/*/arch/2010/001')
+dirs = glob.glob('/data/cosmos2/eng_archive/data/*/arch')
 for dir_ in dirs:
-    files = glob.glob(os.path.join(dir_, '*.fits.gz'))
-    if not files:
-        print 'No files in', dir_
-        continue
+    content = os.path.basename(os.path.split(dir_)[0])
+    print 'Finding units in', dir_
+    for dirpath, dirnames, filenames in os.walk(dir_, topdown=False):
+        files = [f for f in filenames if f.endswith('.fits.gz')]
+        if not files:
+            print 'No fits files in', dirpath
+            continue
 
-    print 'Reading', files[0]
-    hdus = pyfits.open(files[0])
-    cols = hdus[1].columns
-    for msid, unit in zip(cols.names, cols.units):
-        unit = unit.strip()
-        if unit:
-            units[msid.upper()] = unit
-    hdus.close()
+        print 'Reading', files[0]
+        hdus = pyfits.open(os.path.join(dirpath, files[0]))
+        cols = hdus[1].columns
+        for msid, unit in zip(cols.names, cols.units):
+            unit = unit.strip()
+            if unit:
+                msid = msid.upper()
+                if content in ALIASES:
+                    msid = ALIASES[content].get(msid, msid)
+                if re.match(r'(orbit|lunar|solar|angle)ephem', content):
+                    msid = '{}_{}'.format(content.upper(), msid)
+                units[msid.upper()] = unit
+        hdus.close()
+        break
 
 # AFAIK these are the only temperature MSIDs that are actually temperature
 # differences and which require special handling on conversion.
@@ -38,9 +58,14 @@ relative_temp_msids = (
 for msid in relative_temp_msids:
     units[msid] = 'deltaK'
 
+units['3MRMMXMV'] = 'PWM'
+
 # Use info about DEA HK telemetry from converters to add units
-for col in Ska.engarchive.converters._get_deahk_cols():
+for col in _get_deahk_cols():
     if 'unit' in col:
         units[col['name'].upper()] = col['unit']
 
-pickle.dump(units, open('units_cxc.pkl', 'w'))
+if not opt.dry_run:
+    pickle.dump(units, open('units_cxc.pkl', 'w'))
+else:
+    print repr(units)
